refactor: move to proper package structure
